# Conflicts:
#	backend-nest/src/api/persist/persist.service.ts

+++ backend-nest/src/api/classgroups/classgroups.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateClassgroupDto } from './dto/create-classgroup.dto';
 import { PrismaService } from '../../prisma/prisma.service';
 import { classgroup } from '@prisma/client';
@@ -8,12 +8,18 @@ export class ClassgroupsService {
   constructor(private prisma: PrismaService) {}
 
   async create(createClassgroupDto: CreateClassgroupDto) {
+    if (!createClassgroupDto) {
+      throw new Error('Classgroup cannot be null');
+    }
     return this.prisma.classgroup.create({
       data: createClassgroupDto,
     });
   }
 
   async createMany(createClassgroupDtos: CreateClassgroupDto[]) {
+    if (!createClassgroupDtos) {
+      throw new Error('Classgroup cannot be null');
+    }
     return this.prisma.classgroup.createMany({
       data: createClassgroupDtos,
     });
@@ -24,18 +30,33 @@ export class ClassgroupsService {
   }
 
   async findOne(id: number) {
+    const existingClassgroup = await this.prisma.classgroup.findUnique({
+      where: { id },
+    });
+    if (!existingClassgroup) {
+      throw new NotFoundException('Classgroup not found');
+    }
     return this.prisma.classgroup.findUnique({
       where: { id: id },
     });
   }
 
   async findByFile(file: string) {
+    const existingClassgroup = await this.prisma.classgroup.findFirst({
+      where: { file },
+    });
+    if (!existingClassgroup) {
+      throw new NotFoundException('Classgroup not found');
+    }
     return this.prisma.classgroup.findFirst({
       where: { file: file },
     });
   }
 
   async delete(id: number) {
+    if (!id) {
+      throw new Error('The class does not exist');
+    }
     return this.prisma.classgroup.delete({
       where: { id: id },
     });
@@ -46,7 +67,17 @@ export class ClassgroupsService {
     update: Omit<classgroup, 'id'>;
     where: { file: string };
   }) {
-    const isExisting = await this.findByFile(param.where.file);
+    let isExisting;
+    try {
+      isExisting = await this.findByFile(param.where.file);
+    } catch (error) {
+      if (error instanceof NotFoundException) {
+        isExisting = null;
+      } else {
+        throw error;
+      }
+    }
+
     if (isExisting) {
       return this.prisma.classgroup.update({
         where: { id: isExisting.id },
